添加文件

+++ news/app.py
@@ -1,12 +1,12 @@
 from colorama import Fore, Style
 from getpass import getpass
-from news.service.user_service import UserService
+from service.user_service import UserService
 import os
 import sys
 import time
-from news.service.news_service import NewsService
-from news.service.role_service import RoleService
-from news.service.type_service import TypeService
+from service.news_service import NewsService
+from service.role_service import RoleService
+from service.type_service import TypeService
 
 __user_service = UserService()
 __news_service = NewsService()
@@ -59,7 +59,56 @@ while True:
                             print("\n\t保存成功(3秒自动返回)")
                             time.sleep(3)
                     elif opt == '2':
-                        pass
+                        page = 1
+                        while True:
+                            os.system("cls")
+                            count_page = __news_service.search_count_page()
+                            result = __news_service.search_list(page, 10)
+                            for index in range(len(result)):
+                                one = result[index]
+                                print(Fore.LIGHTBLUE_EX,
+                                      "\n\t%d\t%s\t%s\t%s" % (index + 1, one[1], one[2], one[3]))
+                            print(Fore.LIGHTBLUE_EX, "\n\t=====================")
+                            print(Fore.LIGHTBLUE_EX, "\n\t%d/%d" % (page, count_page))
+                            print(Fore.LIGHTBLUE_EX, "\n\t=====================")
+                            print(Fore.LIGHTRED_EX, "\n\tback.返回上一层")
+                            print(Fore.LIGHTRED_EX, "\n\tprev.上一页")
+                            print(Fore.LIGHTRED_EX, "\n\tnext.下一页")
+                            print(Style.RESET_ALL)
+                            opt = input("\n\t输入操作编号：")
+                            if opt == "back":
+                                break
+                            elif opt == "prev" and page > 1:
+                                page -= 1
+                            elif opt == "next" and page < count_page:
+                                page += 1
+                            elif opt.isdigit() and int(opt) >= 1 and int(opt) <= 10:
+                                news_id = result[int(opt) - 1][0]
+                                result = __news_service.search_by_id(news_id)
+                                title = result[0]
+                                type = result[1]
+                                is_top = result[2]
+                                print("\n\t新闻原标题：%s" %(title))
+                                new_title = input("\n\t新标题：")
+                                print("\n\t新闻原类型：%s" % (type))
+                                result = __type_service.search_list()
+                                for index in range(len(result)):
+                                    one = result[index]
+                                    print(Fore.LIGHTBLUE_EX, "\n\t%d.%s" % (index + 1, one[1]))
+                                opt = input("\n\t类型编号：")
+                                type_id = result[int(opt) - 1][0]
+                                # todo 输入新闻内容
+                                content_id = 100
+                                print("\n\t原置顶级别：%s" % (is_top))
+                                new_is_top = input("\n\t置顶级别(0-5):")
+                                is_commite = input("\n\t是否提交(Y/N):")
+                                if is_commite == "Y" or is_commite == "y":
+                                    __news_service.update(news_id, new_title, type_id, content_id, new_is_top)
+                                    print("\n\t保存成功(3秒自动返回)")
+                                    time.sleep(3)
+
+
+
                 elif role == "管理员":
                     print(Fore.LIGHTGREEN_EX, "\n\t1.新闻管理")
                     print(Fore.LIGHTGREEN_EX, "\n\t2.用户管理")
@@ -96,17 +145,27 @@ while True:
                                     print(Fore.LIGHTRED_EX, "\n\tnext.下一页")
                                     print(Style.RESET_ALL)
                                     opt = input("\n\t输入操作编号：")
+                                    # print(opt.isdigit())
                                     if opt == "back":
                                         break
                                     elif opt == "prev" and page > 1:
                                         page -= 1
                                     elif opt == "next" and page < count_page:
                                         page += 1
-                                    elif int(opt) >=1 and int(opt) <= 10:
+                                    elif opt.isdigit() and int(opt) >=1 and int(opt) <= 10:
                                         news_id = result[int(opt)-1][0]
                                         __news_service.update_unreview_news(news_id)
                                         result = __news_service.search_cache(news_id)
-
+                                        title = result[0]
+                                        username = result[1]
+                                        type = result[2]
+                                        content_id = result[3]
+                                        # todo 查找新闻正文
+                                        content = "100"
+                                        is_top = result[4]
+                                        create_time = str(result[5])
+                                        # print(title)
+                                        __news_service.cache_news(news_id, title, username, type, content, is_top, create_time)
                             elif opt == "2":
                                 page = 1
                                 while True:
@@ -131,9 +190,10 @@ while True:
                                         page -= 1
                                     elif opt == "next" and page < count_page:
                                         page += 1
-                                    elif int(opt) >= 1 and int(opt) <= 10:
+                                    elif opt.isdigit() and int(opt) >= 1 and int(opt) <= 10:
                                         news_id = result[int(opt) - 1][0]
                                         __news_service.delete_by_id(news_id)
+                                        __news_service.delete_cache(news_id)
                             else:
                                 break
                     elif opt == '2':
@@ -191,7 +251,7 @@ while True:
                                         page -= 1
                                     elif opt == "next" and page < count_page:
                                         page += 1
-                                    elif int(opt) >= 1 and int(opt) <= 10:
+                                    elif opt.isdigit() and int(opt) >= 1 and int(opt) <= 10:
                                         os.system("cls")
                                         user_id = result[int(opt) - 1][0]
                                         username = input("\n\t新用户名：")
@@ -239,7 +299,7 @@ while True:
                                         page -= 1
                                     elif opt == "next" and page < count_page:
                                         page += 1
-                                    elif int(opt) >= 1 and int(opt) <= 10:
+                                    elif opt.isdigit() and int(opt) >= 1 and int(opt) <= 10:
                                         user_id = result[int(opt) - 1][0]
                                         __user_service.delete_by_id(user_id)
                                         print("\n\t删除成功(3s自动返回)")
